refactor: move functions and adjust tests
