+++ src/utils/string.ts
@@ -1,3 +1,7 @@
 export const capitalize = (input: string) => {
   return input[0].toUpperCase() + input.slice(1);
 };
+
+export const makeIntoSlugs = (stk: string) => {
+  return stk.replaceAll('-', '').toLowerCase();
+};
